backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-dotenv>=1.0.1
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
pydantic>=2.12.0
MetaTrader5>=5.0.5488
